refactor(tools): refactor util functions
