+++ pyvium/util/__init__.py
@@ -1,1 +1,2 @@
 from .get_ivium_dll_path import get_ivium_dll_path
+from .get_file_list import get_file_list
